# ita_root/ita_api_organization/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: ITA Version2.0 API
  description: ITAv2.0用API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute:
    post:
      tags:
      - Driver Controll
      description: Driver作業実行
      operationId: post_driver_excecute
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/driver_execute_body'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_17'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute_dry_run:
    post:
      tags:
      - Driver Controll
      description: |
        Driver作業実行（ドライラン）</br>
        AnsibleDriverの場合はDryRun、TerraformDriverの場合はPlanのみの実行を行う
      operationId: post_driver_execute_dry_run
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/driver_execute_dry_run_body'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_17'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute_check_parameter:
    post:
      tags:
      - Driver Controll
      description: Driver作業実行(実行時に使用するパラメータ確認)
      operationId: post_driver_execute_check_parameter
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/driver_execute_check_parameter_body'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_17'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/{execution_no}:
    get:
      tags:
      - Driver Controll
      description: Driver作業実行の状態取得
      operationId: get_driver_execute_data
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: execution_no
        in: path
        description: 実行No
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/{execution_no}/cancel:
    patch:
      tags:
      - Driver Controll
      description: Driver作業実行の予約取消
      operationId: post_driver_cancel
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: execution_no
        in: path
        description: 実行No
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/{execution_no}/scram:
    patch:
      tags:
      - Driver Controll
      description: Driver作業実行の緊急停止
      operationId: post_driver_scram
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: execution_no
        in: path
        description: 実行No
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute/info:
    get:
      tags:
      - Driver Controll
      description: "Movement,Operationのメニューの基本情報および項目情報を取得する"
      operationId: get_driver_execute_info
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_18'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute/filter/count/{target}:
    post:
      tags:
      - Driver Controll
      description: "Movement,Operationを対象に、検索条件を指定し、レコードの件数する"
      operationId: post_driver_execute_filter_count
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: target
        in: path
        description: movement_list or operation_list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/count_target_body_1'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute/filter/{target}:
    post:
      tags:
      - Driver Controll
      description: "Movement,Operationを対象に、検索条件を指定し、レコードを取得する"
      operationId: post_driver_execute_filter
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: target
        in: path
        description: movement_list or operation_list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/filter_target_body_1'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_5'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/driver/execute/filter/{target}/search/candidates/{column}:
    get:
      tags:
      - Driver Controll
      description: 表示フィルタで利用するプルダウン検索の候補一覧を取得する
      operationId: get_driver_execute_search_candidates
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: target
        in: path
        description: movement_list or operation_list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: column
        in: path
        description: REST用項目名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.driver_controll_controller
components:
  schemas:
...
    ERROR_TEMPLATE:
      type: object
      properties:
        result:
          type: string
          example: 001-002
        message:
          type: string
    FILTER_REQUEST_PARAMETERS:
      type: object
      properties:
        NORMAL:
          type: string
        LIST:
          type: array
          items: {}
        RANGE:
          $ref: '#/components/schemas/FILTER_REQUEST_PARAMETERS_RANGE'
    FILTER_PARAMETERS:
      type: object
      properties:
        parameter:
          $ref: '#/components/schemas/FILTER_PARAMETERS_parameter'
        file:
          $ref: '#/components/schemas/FILTER_PARAMETERS_file'
      example:
        file:
          rest_key_name: base64 encode file
        parameter:
          rest_key_name: rest_key_name
...
    inline_response_200_3:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: array
          example:
            id1: name1
            id2: name2
            id3: name3
          items: {}
        message:
          type: string
      example:
        result: 000-00000
        data:
          id1: name1
          id2: name2
          id3: name3
        message: message
    inline_response_200_4:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: number
        message:
          type: string
      example:
        result: 000-00000
        data: 0.80082819046101150206595775671303272247314453125
        message: message
...
    inline_response_200_5:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: array
          items:
            $ref: '#/components/schemas/FILTER_PARAMETERS'
        message:
          type: string
      example:
        result: 000-00000
        data:
        - file:
            rest_key_name: base64 encode file
          parameter:
            rest_key_name: rest_key_name
        - file:
            rest_key_name: base64 encode file
          parameter:
            rest_key_name: rest_key_name
        message: message
...
    inline_response_200_11:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: boolean
        message:
          type: string
      example:
        result: 000-00000
        data: true
        message: message
...
    driver_execute_body:
      type: object
      properties:
        movement_id:
          type: string
        operation_id:
          type: string
        schedule_date:
          type: string
          example: YYYY/MM/DD hh:mm
      description: 作業実行するDriverの構成情報
    inline_response_200_17:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: object
          example:
            execution_no: string
        message:
          type: string
      example:
        result: 000-00000
        data:
          execution_no: string
        message: message
    driver_execute_dry_run_body:
      type: object
      properties:
        movement_id:
          type: string
        operation_id:
          type: string
        schedule_date:
          type: string
          example: YYYY/MM/DD hh:mm
      description: 作業実行するDriverの構成情報
    driver_execute_check_parameter_body:
      type: object
      properties:
        movement_id:
          type: string
        operation_id:
          type: string
        schedule_date:
          type: string
          example: YYYY/MM/DD hh:mm
      description: 作業実行するDriverの構成情報
    inline_response_200_18:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          $ref: '#/components/schemas/inline_response_200_18_data'
        message:
          type: string
      example:
        result: 000-00000
        data:
          operation_list: null
          movement_list:
            column_group_info:
              key:
                parent_column_group_name: parent_column_group_name
                columns:
                - ""
                - ""
                parent_column_group_id: parent_column_group_id
                column_group_id: column_group_id
                column_group_name: column_group_name
            menu_info:
              sheet_type: sheet_type
              history_table_flag: true
              row_update_flag: true
              web_print_limit: 7
              initial_filter_flg: true
              columns:
              - ""
              - ""
              row_reuse_flag: true
              vertical: true
              row_insert_flag: true
              privilege: 2
              row_disuse_flag: true
              table_name: table_name
              parent_menu_group_name: parent_menu_group_name
              menu_name: menu_name
              menu_group_name: menu_group_name
              auto_filter_flg: true
              menu_id: menu_id
              view_name: view_name
              web_print_confirm: 9
              sort_key: sort_key
              parent_menu_group_id: parent_menu_group_id
              menu_group_id: menu_group_id
              xls_print_limit: 3
              menu_info: menu_info
              inherit: true
              login_necessity: true
            column_info:
              key:
                initial_value: initial_value
                column_disp_seq: 6
                required_item: 2
                ref_table_name: ref_table_name
                save_type: save_type
                column_name: column_name
                unique_item: 5
                description: description
                input_item: 5
                before_validate_register: {}
                col_name: col_name
                column_name_rest: column_name_rest
                column_id: column_id
                ref_multi_lang: ref_multi_lang
                validate_option: {}
                ref_pkey_name: ref_pkey_name
                auto_input: 1
                view_item: view_item
                ref_col_name: ref_col_name
                column_group_id: column_group_id
                after_validate_register: {}
                column_type: 0
                column_group_name: column_group_name
                ref_sort_conditions: ref_sort_conditions
        message: message
    count_target_body_1:
      properties:
        rest_key_name:
          $ref: '#/components/schemas/FILTER_REQUEST_PARAMETERS'
      description: 検索条件を設定
    filter_target_body_1:
      properties:
        rest_key_name:
          $ref: '#/components/schemas/FILTER_REQUEST_PARAMETERS'
      description: 検索条件を設定
    FILTER_REQUEST_PARAMETERS_RANGE:
      type: object
      properties:
        START:
          type: string
        END:
          type: string
      example: null
    FILTER_PARAMETERS_parameter:
      type: object
      properties:
        rest_key_name:
          type: string
      example:
        rest_key_name: rest_key_name
    FILTER_PARAMETERS_file:
      type: object
      properties:
        rest_key_name:
          type: string
          example: base64 encode file
      example:
        rest_key_name: base64 encode file
...
    inline_response_200_data:
      type: object
      properties:
        column_group_info:
          $ref: '#/components/schemas/COLUMN_GROUP_INFO'
        column_info:
          $ref: '#/components/schemas/COLUMN_INFO'
        menu_info:
          $ref: '#/components/schemas/MENU_INFO'
      example:
        column_group_info:
          key:
            parent_column_group_name: parent_column_group_name
            columns:
            - ""
            - ""
            parent_column_group_id: parent_column_group_id
            column_group_id: column_group_id
            column_group_name: column_group_name
        menu_info:
          sheet_type: sheet_type
          history_table_flag: true
          row_update_flag: true
          web_print_limit: 7
          initial_filter_flg: true
          columns:
          - ""
          - ""
          row_reuse_flag: true
          vertical: true
          row_insert_flag: true
          privilege: 2
          row_disuse_flag: true
          table_name: table_name
          parent_menu_group_name: parent_menu_group_name
          menu_name: menu_name
          menu_group_name: menu_group_name
          auto_filter_flg: true
          menu_id: menu_id
          view_name: view_name
          web_print_confirm: 9
          sort_key: sort_key
          parent_menu_group_id: parent_menu_group_id
          menu_group_id: menu_group_id
          xls_print_limit: 3
          menu_info: menu_info
          inherit: true
          login_necessity: true
        column_info:
          key:
            initial_value: initial_value
            column_disp_seq: 6
            required_item: 2
            ref_table_name: ref_table_name
            save_type: save_type
            column_name: column_name
            unique_item: 5
            description: description
            input_item: 5
            before_validate_register: {}
            col_name: col_name
            column_name_rest: column_name_rest
            column_id: column_id
            ref_multi_lang: ref_multi_lang
            validate_option: {}
            ref_pkey_name: ref_pkey_name
            auto_input: 1
            view_item: view_item
            ref_col_name: ref_col_name
            column_group_id: column_group_id
            after_validate_register: {}
            column_type: 0
            column_group_name: column_group_name
            ref_sort_conditions: ref_sort_conditions
...
    inline_response_200_18_data:
      type: object
      properties:
        movement_list:
          $ref: '#/components/schemas/inline_response_200_data'
        operation_list:
          $ref: '#/components/schemas/inline_response_200_data'
      example:
        operation_list: null
        movement_list:
          column_group_info:
            key:
              parent_column_group_name: parent_column_group_name
              columns:
              - ""
              - ""
              parent_column_group_id: parent_column_group_id
              column_group_id: column_group_id
              column_group_name: column_group_name
          menu_info:
            sheet_type: sheet_type
            history_table_flag: true
            row_update_flag: true
            web_print_limit: 7
            initial_filter_flg: true
            columns:
            - ""
            - ""
            row_reuse_flag: true
            vertical: true
            row_insert_flag: true
            privilege: 2
            row_disuse_flag: true
            table_name: table_name
            parent_menu_group_name: parent_menu_group_name
            menu_name: menu_name
            menu_group_name: menu_group_name
            auto_filter_flg: true
            menu_id: menu_id
            view_name: view_name
            web_print_confirm: 9
            sort_key: sort_key
            parent_menu_group_id: parent_menu_group_id
            menu_group_id: menu_group_id
            xls_print_limit: 3
            menu_info: menu_info
            inherit: true
            login_necessity: true
          column_info:
            key:
              initial_value: initial_value
              column_disp_seq: 6
              required_item: 2
              ref_table_name: ref_table_name
              save_type: save_type
              column_name: column_name
              unique_item: 5
              description: description
              input_item: 5
              before_validate_register: {}
              col_name: col_name
              column_name_rest: column_name_rest
              column_id: column_id
              ref_multi_lang: ref_multi_lang
              validate_option: {}
              ref_pkey_name: ref_pkey_name
              auto_input: 1
              view_item: view_item
              ref_col_name: ref_col_name
              column_group_id: column_group_id
              after_validate_register: {}
              column_type: 0
              column_group_name: column_group_name
              ref_sort_conditions: ref_sort_conditions